pyproject.toml
[tool.poetry]
name = "polisci-427s"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-igraph = "^0.9.6"
pycairo = "^1.20.1"
matplotlib = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
